feat(material): move types into separate entry points
