+++ src/material/checkbox/src/public_api.ts
@@ -1,0 +1,3 @@
+export { FormlyMatCheckboxModule } from './checkbox.module';
+export { FormlyFieldCheckbox } from './checkbox.type';
+export { FormlyFieldMultiCheckbox } from './multicheckbox.type';
