+++ src/material/datepicker/src/datepicker.module.ts
@@ -2,7 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormlyModule } from '@ngx-formly/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import { MatFormFieldModule } from '@angular/material/form-field';
+
+import { FormlyMatFormFieldModule } from '@ngx-formly/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 
@@ -13,9 +14,10 @@ import { FormlyDatepickerTypeComponent } from './datepicker.type';
   imports: [
     CommonModule,
     ReactiveFormsModule,
-    MatFormFieldModule,
     MatInputModule,
     MatDatepickerModule,
+
+    FormlyMatFormFieldModule,
     FormlyModule.forChild({
       types: [{
         name: 'datepicker',
