+++ src/material/datepicker/src/datepicker.type.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { FieldType } from '@ngx-formly/material';
+import { FieldType } from '@ngx-formly/material/form-field';
 import { MatInput } from '@angular/material/input';
 
 @Component({
