+++ src/material/form-field/package.json
@@ -1,0 +1,3 @@
+{
+  "ngPackage": {}
+}
