+++ src/material/form-field/src/field.type.ts
@@ -2,7 +2,7 @@ import { OnInit, OnDestroy, AfterViewInit, TemplateRef, ViewChild } from '@angul
 import { FieldType as CoreFieldType } from '@ngx-formly/core';
 import { Subject } from 'rxjs';
 import { MatFormField, MatFormFieldControl } from '@angular/material/form-field';
-import { FormlyErrorStateMatcher } from '../formly.error-state-matcher';
+import { FormlyErrorStateMatcher } from './formly.error-state-matcher';
 
 export abstract class FieldType extends CoreFieldType implements OnInit, AfterViewInit, OnDestroy, MatFormFieldControl<any> {
   @ViewChild('matPrefix') matPrefix: TemplateRef<any>;
