+++ src/material/form-field/src/public_api.ts
@@ -1,0 +1,2 @@
+export { FormlyMatFormFieldModule } from './form-field.module';
+export { FieldType } from './field.type';
