+++ src/material/input/src/public_api.ts
@@ -1,0 +1,3 @@
+export { FormlyMatInputModule } from './input.module';
+export { FormlyFieldInput } from './input.type';
+export { FormlyFieldTextArea } from './textarea.type';
