+++ src/material/ng-package.json
@@ -4,7 +4,12 @@
   "deleteDestPath": false,
   "lib": {
     "entryFile": "src/public_api.ts",
+    "form-field": [{ "path": "form-field" }],
+    "input": [{ "path": "input" }],
+    "select": [{ "path": "select" }],
+    "checkbox": [{ "path": "checkbox" }],
     "datepicker": [{ "path": "datepicker" }],
+    "radio": [{ "path": "radio" }],
     "toggle": [{ "path": "toggle" }],
     "slider": [{ "path": "slider" }],
     "externals": {
