+++ src/material/radio/index.ts
@@ -1,0 +1,1 @@
+export * from './src/public_api';
