+++ src/material/radio/package.json
@@ -1,0 +1,3 @@
+{
+  "ngPackage": {}
+}
