+++ src/material/radio/src/public_api.ts
@@ -1,0 +1,2 @@
+export { FormlyMatRadioModule } from './radio.module';
+export { FormlyFieldRadio } from './radio.type';
