+++ src/material/select/src/public_api.ts
@@ -1,0 +1,2 @@
+export { FormlyMatSelectModule } from './select.module';
+export { FormlyFieldSelect } from './select.type';
