+++ src/material/select/src/select.module.ts
@@ -2,31 +2,33 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormlyModule } from '@ngx-formly/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import { MatSliderModule } from '@angular/material/slider';
-import { FormlyMatFormFieldModule } from '@ngx-formly/material/form-field';
 
-import { FormlySliderTypeComponent } from './slider.type';
+import { FormlyMatFormFieldModule } from '@ngx-formly/material/form-field';
+import { MatSelectModule } from '@angular/material/select';
+
+import { FormlyFieldSelect } from './select.type';
+import { FormlySelectOptionsPipe } from './select-options.pipe';
 
 @NgModule({
-  declarations: [FormlySliderTypeComponent],
+  declarations: [FormlyFieldSelect, FormlySelectOptionsPipe],
   imports: [
     CommonModule,
     ReactiveFormsModule,
-    MatSliderModule,
+    MatSelectModule,
+
     FormlyMatFormFieldModule,
     FormlyModule.forChild({
       types: [{
-        name: 'slider',
-        component: FormlySliderTypeComponent,
+        name: 'select',
+        component: FormlyFieldSelect,
         wrappers: ['form-field'],
         defaultOptions: {
           templateOptions: {
-            hideFieldUnderline: true,
-            floatLabel: 'always',
+            options: [],
           },
         },
       }],
     }),
   ],
 })
-export class FormlyMatSliderModule { }
+export class FormlyMatSelectModule { }
