--- src/material/src/lib/types/types.ts
@@ -1,7 +1,0 @@
-export { FieldType } from './field';
-export { FormlyFieldCheckbox } from './checkbox';
-export { FormlyFieldMultiCheckbox } from './multicheckbox';
-export { FormlyFieldInput } from './input';
-export { FormlyFieldRadio } from './radio';
-export { FormlyFieldTextArea } from './textarea';
-export { FormlyFieldSelect } from './select';
