--- src/material/src/lib/wrappers/wrappers.ts
@@ -1,1 +1,0 @@
-export { FormlyWrapperFormField } from './form-field';
