+++ src/material/src/test.ts
@@ -17,6 +17,6 @@ getTestBed().initTestEnvironment(
   platformBrowserDynamicTesting(),
 );
 // Then we find all the tests.
-const context = require.context('./', true, /\.spec\.ts$/);
+const context = require.context('../', true, /\.spec\.ts$/);
 // And load the modules.
 context.keys().map(context);
