+++ src/material/toggle/src/toggle.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormlyModule } from '@ngx-formly/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { FormlyMatFormFieldModule } from '@ngx-formly/material/form-field';
 
 import { FormlyToggleTypeComponent } from './toggle.type';
 
@@ -12,6 +13,7 @@ import { FormlyToggleTypeComponent } from './toggle.type';
     CommonModule,
     ReactiveFormsModule,
     MatSlideToggleModule,
+    FormlyMatFormFieldModule,
     FormlyModule.forChild({
       types: [{
         name: 'toggle',
